Make options not being repeated in the elements definitions
