+++ lib/AutoDemand/Filter.ex
@@ -1,19 +1,10 @@
-defmodule AutoDemand.Elements.Filter do
-  use Membrane.Filter
+defmodule AutoDemand.Filter do
+  use Base.Filter
 
   def_input_pad :input, demand_mode: :auto, caps: :any
   def_output_pad :output, demand_mode: :auto, caps: :any
 
-  def_options id: [
-                type: :integer,
-                spec: pos_integer,
-                description: "Id of the element in the pipeline"
-              ],
-              reductions: [
-                type: :integer,
-                spec: pos_integer,
-                description: "Number of reductions which should be done while processing each buffer"
-              ]
+  def_options_with_default()
 
   @impl true
   def handle_init(opts) do
